# backend/docs/swagger.yaml
basePath: /admin-api
definitions:
  github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response:
    properties:
      code:
        example: 0
        type: integer
      data: {}
      msg:
        example: ""
        type: string
    type: object
...
  internal_modules_system.AssignRoleDataScopeRequest:
    properties:
      dataScope:
        type: integer
      dataScopeDeptIds:
        items:
          type: integer
        type: array
      roleId:
        type: integer
    required:
    - dataScope
    - roleId
    type: object
  internal_modules_system.AssignRoleMenuRequest:
    properties:
      menuIds:
        items:
          type: integer
        type: array
      roleId:
        type: integer
    required:
    - roleId
    type: object
  internal_modules_system.AssignUserRoleRequest:
    properties:
      roleIds:
        items:
          type: integer
        type: array
      userId:
        type: integer
    required:
    - userId
    type: object
...
  internal_modules_system.RoleSaveRequest:
    properties:
      code:
        type: string
      dataScope:
        type: integer
      dataScopeDeptIds:
        items:
          type: integer
        type: array
      id:
        type: integer
      name:
        type: string
      remark:
        type: string
      sort:
        type: integer
      status:
        type: integer
    required:
    - code
    - name
    type: object
...
info:
  contact: {}
  description: 'Go modular-monolith scaffold for operations platforms. Default database:
    MySQL 8.4.'
  title: Nimbus Framework Go API
  version: "1.0"
paths:
...
  /system/permission/assign-role-data-scope:
    post:
      consumes:
      - application/json
      parameters:
      - description: Role data scope
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_system.AssignRoleDataScopeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Assign data scope to a role
      tags:
      - System Permission
  /system/permission/assign-role-menu:
    post:
      consumes:
      - application/json
      parameters:
      - description: Role menu assignment
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_system.AssignRoleMenuRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Assign menus to a role
      tags:
      - System Permission
  /system/permission/assign-user-role:
    post:
      consumes:
      - application/json
      parameters:
      - description: User role assignment
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_system.AssignUserRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Assign roles to an operations-console user
      tags:
      - System Permission
  /system/permission/list-role-menus:
    get:
      parameters:
      - description: Role ID
        in: query
        name: roleId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: List menu IDs assigned to a role
      tags:
      - System Permission
  /system/permission/list-user-roles:
    get:
      parameters:
      - description: User ID
        in: query
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: List role IDs assigned to a user
      tags:
      - System Permission
...
  /system/role/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Role
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_system.RoleSaveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Create a role
      tags:
      - System Role
  /system/role/delete:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Delete a role
      tags:
      - System Role
  /system/role/delete-list:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Delete roles in batch
      tags:
      - System Role
...
  /system/role/get:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Get a role
      tags:
      - System Role
  /system/role/page:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Page roles
      tags:
      - System Role
  /system/role/simple-list:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: List enabled roles
      tags:
      - System Role
  /system/role/update:
    put:
      consumes:
      - application/json
      parameters:
      - description: Role
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_system.RoleSaveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Update a role
      tags:
      - System Role
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"